fixed mod function of #expr not truncating its parameters to integers

+++ info.bliki.wiki/bliki-core/src/main/java/info/bliki/wiki/template/expr/eval/DoubleEvaluator.java
@@ -96,7 +96,7 @@ public class DoubleEvaluator {
 			if (Math.abs(arg2 - 0.0) < DoubleEvaluator.EPSILON) {
 				throw new ArithmeticException("Division by zero");
 			}
-			return arg1 % arg2;
+			return ((int) arg1) % ((int) arg2);
 		}
 	}
